[feat-#30] dto에 기본 생성자 제거
