+++ src/main/java/com/bss/bssserverapi/domain/auth/dto/request/LoginUserReqDto.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor(access = AccessLevel.PACKAGE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class LoginUserReqDto {
 
     @NotNull
