+++ src/main/java/com/bss/bssserverapi/domain/comment/dto/res/GetCommentResDto.java
@@ -1,38 +1,35 @@
-package com.bss.bssserverapi.domain.comment.dto;
+package com.bss.bssserverapi.domain.comment.dto.res;
 
 import com.bss.bssserverapi.domain.comment.Comment;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 
 @Getter
-public class GetReplyCommentResDto {
+@AllArgsConstructor
+public class GetCommentResDto {
 
     private Long id;
     private String content;
     private String userName;
+    private Long childCommentCount;
+    private Long recommendCount;
     private Boolean isDeleted;
     private LocalDateTime deletedAt;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 
-    public GetReplyCommentResDto(final Long id, final String content, final String userName, final Boolean isDeleted, final LocalDateTime deletedAt, final LocalDateTime createdAt, final LocalDateTime updatedAt) {
+    static public GetCommentResDto toDto(final Comment comment) {
 
-        this.id = id;
-        this.content = content;
-        this.userName = userName;
-        this.isDeleted = isDeleted;
-        this.deletedAt = deletedAt;
-        this.createdAt = createdAt;
-        this.updatedAt = updatedAt;
-    }
-
-    static public GetReplyCommentResDto toDto(final Comment comment) {
-
-        return new GetReplyCommentResDto(
+        return new GetCommentResDto(
                 comment.getId(),
                 (comment.getIsDeleted().equals(Boolean.TRUE) ? "삭제된 댓글입니다. " : comment.getContent()),
                 comment.getUser().getUserName(),
+                comment.getChildCommentCount(),
+                comment.getRecommendCount(),
                 comment.getIsDeleted(),
                 comment.getDeletedAt(),
                 comment.getCreatedAt(),
