+++ src/main/java/com/bss/bssserverapi/domain/research/dto/res/GetResearchResDto.java
@@ -1,11 +1,9 @@
-package com.bss.bssserverapi.domain.research.dto;
+package com.bss.bssserverapi.domain.research.dto.res;
 
 import com.bss.bssserverapi.domain.research.Research;
 import com.bss.bssserverapi.domain.tag.Tag;
-import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
