+++ src/main/java/com/bss/bssserverapi/domain/research/service/ResearchService.java
@@ -3,10 +3,10 @@ package com.bss.bssserverapi.domain.research.service;
 import com.bss.bssserverapi.domain.research.Research;
 import com.bss.bssserverapi.domain.research.ResearchRecommend;
 import com.bss.bssserverapi.domain.research.ResearchTag;
-import com.bss.bssserverapi.domain.research.dto.CreateResearchReqDto;
-import com.bss.bssserverapi.domain.research.dto.GetResearchPagingResDto;
-import com.bss.bssserverapi.domain.research.dto.GetResearchPreviewResDto;
-import com.bss.bssserverapi.domain.research.dto.GetResearchResDto;
+import com.bss.bssserverapi.domain.research.dto.req.CreateResearchReqDto;
+import com.bss.bssserverapi.domain.research.dto.res.GetResearchPagingResDto;
+import com.bss.bssserverapi.domain.research.dto.res.GetResearchPreviewResDto;
+import com.bss.bssserverapi.domain.research.dto.res.GetResearchResDto;
 import com.bss.bssserverapi.domain.research.repository.ResearchJpaRepository;
 import com.bss.bssserverapi.domain.research.repository.ResearchRecommendJpaRepository;
 import com.bss.bssserverapi.domain.research.repository.ResearchTagRepository;
