+++ src/main/java/com/bss/bssserverapi/domain/stock/dto/GetStockResDto.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor(access = AccessLevel.PACKAGE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class GetStockResDto {
 
     private String stockCode;
