+++ src/main/java/com/bss/bssserverapi/domain/user/dto/GetUserResDto.java
@@ -1,12 +1,9 @@
 package com.bss.bssserverapi.domain.user.dto;
 
-import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor(access = AccessLevel.PACKAGE)
 public class GetUserResDto {
 
     private String userName;
