+++ src/test/java/com/bss/bssserverapi/research/ResearchServiceTest.java
@@ -3,9 +3,9 @@ package com.bss.bssserverapi.research;
 import com.bss.bssserverapi.domain.research.Research;
 import com.bss.bssserverapi.domain.research.ResearchRecommend;
 import com.bss.bssserverapi.domain.research.ResearchTag;
-import com.bss.bssserverapi.domain.research.dto.CreateResearchReqDto;
-import com.bss.bssserverapi.domain.research.dto.GetResearchPagingResDto;
-import com.bss.bssserverapi.domain.research.dto.GetResearchResDto;
+import com.bss.bssserverapi.domain.research.dto.req.CreateResearchReqDto;
+import com.bss.bssserverapi.domain.research.dto.res.GetResearchPagingResDto;
+import com.bss.bssserverapi.domain.research.dto.res.GetResearchResDto;
 import com.bss.bssserverapi.domain.research.repository.ResearchJpaRepository;
 import com.bss.bssserverapi.domain.research.repository.ResearchRecommendJpaRepository;
 import com.bss.bssserverapi.domain.research.repository.ResearchTagRepository;
